First Funspark send Message

+++ emtours-spring-boot/src/main/java/org/camunda/bpm/emtours/CustomerRequestRepository.java
@@ -2,5 +2,5 @@ package org.camunda.bpm.emtours;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRequestRepository extends CrudRepository<CustomerRequest, Long> {
+public interface CustomerRequestRepository extends CrudRepository<CustomerRequest, Integer> {
 }
