+++ emtours-spring-boot/src/main/java/org/camunda/bpm/emtours/RecommendationRepository.java
@@ -3,5 +3,5 @@ package org.camunda.bpm.emtours;
 import org.camunda.bpm.entities.Recommendation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RecommendationRepository extends CrudRepository<Recommendation, Long> {
+public interface RecommendationRepository extends CrudRepository<Recommendation, Integer> {
 }
